Add rudimentary /teams/:slug action
For now this is just a blank page that shows the team title. I want
something to link to so that I can DRY up the table rows in the
teams/index and authors/index templates, by making them substantially
the same.

In a little bit I'll put some actual content on this page.

Change-Id: I87ad2f4fa96096ae5cc68049196aea9960138c21
Reviewed-on: https://gerrit.causes.com/25847
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>