Propagate "days" param through to Ajax requests
We DRY up some of the controller actions here by handling "days" in a
before filter.

It's pretty clear that as the number of routing helpers grows, we'll
want to split them out into a separate file, but leaving them where they
are for now.

Note that it's necessary to use `::URI` here to get at the top-level
constant, and not `URI`, which in Sinatra context refers to an instance
of URI::Parse.

Change-Id: I5626e4793a73bdd9e8417b60a1944f78a51aa6d9
Reviewed-on: https://gerrit.causes.com/26047
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>