Add a sticky footer and show a random quip in it
As an "easter egg", put a random quip in the footer, drawn from a
collection of programming quotes and aphorisms found pulled from a
handful of Google search results.

The footer is "sticky" in the sense that it stays at the bottom of the
viewport, even if the HTML content itself is short. Note that it is not
a fixed position element; it will move down if the content is long. The
method employed here is the one recommended in the Twitter bootstrap
samples:

  http://twitter.github.io/bootstrap/examples/sticky-footer.html

On small screens, the text ellipsizes, but we show the full quote via a
title attribute.

Change-Id: I71b5a05d02c349c214ac23fa9ec378dfd365b66d
Reviewed-on: https://gerrit.causes.com/26048
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>