Add RepoSet class and use it in Gatherer class
I am about to add some functionality that needs to iterate over the
repos on disk. As this is exactly what the Gatherer class is already
doing, extract that functionality out into a new class called RepoSet,
whose sole responsibility is to provide an abstraction around a set of
repos on disk.

Change-Id: Ib6fd17cd2ec48cb657e8da6486d2323eb215e524
Reviewed-on: https://gerrit.causes.com/26056
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>