Add `git` subcommand, with `git gc` subcommand
We now have a `twigg git gc` sub-subcommand that we can use to run `git
gc` on all the repos in batch mode.

This required me to make the `RepoSet#repos` method public, which I had
initially marked as private.

The command implementation inherits from GitHost in order to leverage
its functionality, which is a bit of a smell because there is little
that is "host-like" about this command (as a stretch argument, you could
allege that your local filesystem is the "host").

For now, this works and is DRY, but I'd like to clean it up conceptually
in a future commit.

Change-Id: I874d274767b7a7ae86644003b6e5b2fe948cfa80
Reviewed-on: https://gerrit.causes.com/26057
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>