Make text more readable on dark backgrounds on /pairs
Once the background crosses a threshold of 0.4 opacity, switch to
off-white text (from black) for better readability. We only do this on
the "solo" cells (the ones with blue backgrounds) because the green
cells are easy to read at any opacity.

The color used, #f8f8ff, is known as "ghost white":

  http://en.wikipedia.org/wiki/Shades_of_white

Change-Id: I3582a956b4bd4ea488c2ff33f5515d638baa73ff
Reviewed-on: https://gerrit.causes.com/26058
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>