Fix bug in pairing matrix max calculations
We were grabbing the max count of the outer loop, but not the nested
inner loop. This returned the right result only some of the time.

Evidently the code could be DRYed up a little, as the two max
calculations are substantially similar. I'll see if I can come up with
something that is not too clever; for now we'' stick with the
dumb-but-readable option.

Change-Id: I305cca9fe8f8169f1b20623e8037e6ff0897843e
Reviewed-on: https://gerrit.causes.com/26059
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>