Get /teams/:slug to parity with /authors/:slug
Here we break the bulk of the /authors/:slug page out into a partial
which can be reused directly by the /teams/:slug page. This means that
the teams detail page now has commit summary, nvd3 visualization, and
commit listing.

This is all possible because it rests on the shared underlying
abstraction provided by the CommitSet class.

In order to facilitate this, moved the d3/nvd3 JS up into the layout.

Change-Id: I199052a967258f1807661fe634d81310eda5fd2a
Reviewed-on: https://gerrit.causes.com/26065
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>