Remove Shotgun in favor of Sinatra::Reloader
I found Shotgun to be mostly unsatisfactory for development. I am not
sure whether it is a problem with the gem itself, or I was expecting too
much of it in the context of Twigg, where the source code is distributed
across multiple directories, but it failed to reload the code for most
changes (view changes worked, although I think they would have worked
anyway, but almost nothing else did, requiring a manual restart).

Rip it out in favor of Sinatra::Reloader, which comes with
sinatra-contrib, which we already depend on. As a bonus, this reduces
our dependency footprint, in addition to allowing us to configure things
so that our source code always gets reloaded, irrespective of which
directory it happens to have come from.

Change-Id: Ibe3120b303c82726410aec5e512870c53c7ce9d1
Reviewed-on: https://gerrit.causes.com/26066
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>