Add "time" icon for switching day counts
With this commit, anywhere we show a day count, we now show a "time"
icon that can be clicked to reveal a Boostrap popover containing
controls for switching the time interval to one of a number of preset
values (3, 7, 14, 30, 60 and 90 days).

As the content is in some places being used inside an h1 or h2 element,
and at others inside the title element in the head of the document, we
add a `strip_tags` helper that can be used to strip out HTML tags before
inserting this content into the title element. This swells the
dependency footprint of the twigg-app gem by pulling in Nokogiri, but
that seemed better than engaging in the hapless task of
parsing/sanitizing HTML with regular expressions.

Note that we can't just throw our `require 'nokogiri'` above the
declaration of the Twigg::App class, because that file gets pulled in
and evaluated by the gemspec, which produces a chicken-and-egg problem
(you can't run `bundle install` or even `bundle check` because you can't
parse the gemspec without exploding). There might be a better solution
for this but for now that will do.

Change-Id: Ife006b3d2f7da8a51575e365afc8f74d8efd4feb
Reviewed-on: https://gerrit.causes.com/26069
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>