Add regression specs for PairMatrix class
This commit adds specs that test the commit-counting logic of the
PairMatrix class; the specific motivation here is to prevent a
regression of the bug fixed in commit 3be3791, where we were
double-counting solo commits.

As we don't need a real repo to exercise this code, we set up a Factory
Girl factory for producing Commit instances. For the time-being, we
provide those Commit instances with a phoney "Repo" fake (an OpenStruct
that implements the one method that the Commit class cares about,
`#name`).

Included are sequences for generating realistic-looking Commit objects,
including names (the result of some Google searching and some massaging
with awk, Ruby and Vim) and subjects (sampled from the Twigg repo
itself). We use unique sequences where possible/practical, and random
ones where it simply doesn't matter. We're not currently making much use
of this fanciness, but I did want the returned objects to look like real
commits.

Change-Id: Iaf38aacb05c81c57db19bb94342df9c446dd97a9
Reviewed-on: https://gerrit.causes.com/26186
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>