Add scaffolding for /gerrit path
This commit creates a space into which we will insert Gerrit stats (the
equivalent of the `twigg gerrit stats` command-line command. Here we add
a route and a template and a basic controller action. At the moment it
basically renders an empty page. I'll follow up shortly with a commit
(or two) that builds out the back-end and plugs it in to the front-end.

Note that this feature won't be enabled by default; you have to
explicitly turn it on. The reason is that most places that use Twigg
most likely won't be using Gerrit. We could also try to be smart and
assume that you want to use /gerrit if you have the twigg-gerrit gem
installed, but let's start dumb for now.

Change-Id: Ia5020d0ff9803c8983f97a606f8794edc26fe5cc
Reviewed-on: https://gerrit.causes.com/26190
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>