Move `#age` method into `Twigg::Util`
This never really belonged in the `Twigg::Command::Gerrit` class; it was
just there for expediency during the development of the `twigg gerrit
stats` subcommand.

As we're about to expose the Gerrit stats in the web app, we need to
scoop some of the functionality out of the command class and put it
where it will also be useful to the web app.

This commit is a first step in that direction.

Change-Id: Idcf1a44d746008ceca4225f5deb093437df00d0c
Reviewed-on: https://gerrit.causes.com/26191
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>