Move (most of) Gerrit database interaction into separate class
This commit moves the core database interaction out of the Gerrit
command class and into a dedicated DB class. This required some
rejigging of the `with_dependency` method to make it accessible from the
various places that we now need it.

A subsequent commit will move the remaining DB interactions into a
better location, in service of getting the Gerrit stats into the web
app.

Change-Id: I5e43d21c56e0e3729775a6115296e1794b37c556
Reviewed-on: https://gerrit.causes.com/26192
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>