Display open changes at /gerrit
This will work only in the development environment for now, because it
assumes the presence of the twigg-gerrit gem.

Some reorganizing will be required (in future commits) so that we can
use `#with_dependency` to bring in the gem and degrade with an
informative diagnostic message when it is not present.

Change-Id: I5e7fb57cae2d101bc8750f29b914602341d2ea13
Reviewed-on: https://gerrit.causes.com/26194
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>