Move App server to App::Server
Having the App class serve simultaneously as the "namespace" for the
twigg-app gem and also the Sinatra::Base subclass that we use to run the
web app was making it difficult to use `#with_dependency` in the app.

Specifically, the App class was being loaded at gemspec-evaluation time
(in order to access the VERSION constant), which meant that we couldn't
use things like `#with_dependency` in any meaninfgul way (doing so would
occur before the gem itself had been fully loaded, creating a circular
dependency, because `#with_dependency` is defined in the twigg gem, and
the twigg gem depends on the twigg-app gem in development).

This commit makes the problem go away by cleaning out App so that it
serves only as a namespace, and moving the server code into a new class,
App::Server. Now everything is just dandy.

Change-Id: Ie5720a24dd84d6a5e54f7c67ea6dfd513ba2f889
Reviewed-on: https://gerrit.causes.com/26196
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>