Add `#with_dependency` call to Twigg::App::Server
With this change, Twigg will warn you (or rather, explode loudly) if you
try to boot the server with Gerrit support turned on but without the
twigg-gerrit gem installed.

In order to get `#with_dependency` to work in this context we needed to
make the `Console.die` method public, so that it could be called from
anywhere. The alternative of having `#with_dependency` require the
Console methods to have been mixed in was deemed too subtle and flakey,
as was the thought of bundling `#with_dependency` into Console itself,
where it doesn't seem to belong.

Change-Id: Ieb6080c141a39977ca0d9cb8c981ea305b9b1a4a
Reviewed-on: https://gerrit.causes.com/26197
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>