Add .ruby-version
We're already locking down development-side dependencies with Bundler
and the Gemfile.lock, so we may as well do the same for the version of
Ruby that we use too.

We wouldn't want to be this prescriptive with the shipped version of the
gem, but for development purposes, this is convenient.

Change-Id: I3d0e20118ecc973198366a2753cc1cde322ec014
Reviewed-on: https://gerrit.causes.com/26206
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>