Add Russian novel easter egg
Accessible at /russian-novels, or via the "я" link in the global nav
bar, this page shows a bubble chart representing the Russianness of an
author's commit messages over the requested interval.

  - bubble size indicates overall Russianness (commit message length)
  - bubble text shows the author name
  - bubble hover shows detailed stats on Russianness, Flesch Reading
    Ease score (lower scores indicate harder-to-read commit messages)
    and author and team name
  - bubble color indicates team membership

This is a quick easter egg hacked together for fun, so I will probably
loop back later and clean up the overall placement of the code (as an
easter egg, it perhaps warrants being better encapsulated off somewhere
peripheral, rather than cluttering up the main Commit and CommitSet
classes).

Change-Id: Iaf7c374f913ed271fc16c674a68386c8126634ee
Reviewed-on: https://gerrit.causes.com/26616
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>