Resize bubble chart to match window
Make the bubble chart as large as possible. This means that we try to
make the SVG fill up as much width as the `#bubble-chart` div has
available, but we don't ever go any wider than `window.innerHeight *
1.2`, because on very wide displays this would induce an awkward need to
scroll in order to see the entire chart.

Additionally, to make things look nice, center the chart.

Note that the dimensions are calculated once only, when the window is
first displayed. We don't make any attempt to recalculate on window
resize.

Change-Id: I6c30ab2b100fc3e6502982227a15695abdf03dda
Reviewed-on: https://gerrit.causes.com/26619
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>