Display progress bar while loading bubble-chart
We are listening for progress events, but without streaming on the
server we're not really going to get any of those in a meaningful way
(especially because we can't start streaming the JSON until the entire
commit set machinery has finished its job).

So, we just throw up a crude indeterminate progress bar.

It's better than nothing.

Change-Id: Icb9d2b51f23013d1aa451254ed65d4e7129b8ed9
Reviewed-on: https://gerrit.causes.com/26623
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>