Start using Bower to manage twigg-app dependencies
This commit switches over to Bower for all of twigg-app's front-end
dependencies, with the exception of replace.js, whose Bower-ification is
currently awaiting review here:

  https://gerrit.causes.com/26425

This means that d3, nvd3, jquery, bootstrap, and boostrap-glyphicons are
all being pulled down into the "bower_components" directory with `bower
install`. This will make updating easier than manually downloading and
placing new versions.

For now, replace.js has been manually copied over.

Note that `gem build` does the wrong thing if "public/vendor" is just a symlink to
"bower_components (it adds only normal files to the gem archive). We can
work around this by changing the direction of the symlink:
"public/vendor" is the real directory and "bower_components" is just a
symlink to it; `bower install` does the right thing.

One other minor thing to note: the bootstrap-glyphicons project doesn't
have any tags right now, so we have to point the "version" at
"gh-pages", which is the only branch in the repo.

Change-Id: I5eef3454cd04e067ea6b20a4dfebd83811e9a4e3
Reviewed-on: https://gerrit.causes.com/26627
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>