Remove dependency on nvd3
Replace the nvd3 bar chart with a pure-d3 implementation. This reduces
our dependency footprint and gives us finer-grained control over how
these things actually look. This first cut is a little crude, but it is
close to parity with what we had before. In developing this I took
pointers from:

  http://bl.ocks.org/mbostock/3885304
  http://jsfiddle.net/chrisJamesC/MmEjF/34/

Change-Id: Iac55f07737cd4a7b0ba22639037f4ffa4c5b48d9
Reviewed-on: https://gerrit.causes.com/26632
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>