Show tooltips on commit bar chart hover
Show the commit count and the date; this is useful on rows where the bar
itself is too small to show a number, or there are two many bars to
distinguish the dates easily.

Note that we need to tell Bootstrap where to place the tooltips, or it
will do its default behavior and insert a div in the middle of the svg,
which the browser won't like.

Change-Id: I1bbdbb8987fb21d2f2c4e2d7d3db4619bd64b3a9
Reviewed-on: https://gerrit.causes.com/26634
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>