Further DRY up help subcommand emission
This makes it even easier to add help for new subcommands.

Along the way I made the output format more compact (no blank lines
under topic headers, and no indenting); this was initially just going to
be an intermediate state on the way to maintaining parity with what the
output looked like before, but on inspecting it while testing, I found
it to look better so am going to leave it.

Change-Id: I844fb3f6ccc0157e509279502017dfaead8b95cb
Reviewed-on: https://gerrit.causes.com/26638
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>