Teach Flesch class to never return NaN values
JSON doesn't allow NaN values, and empty commit messages produce exactly
that. (`0 / 0` raises a ZeroDivisionError, but `0 / 0.0` simply returns
`NaN`).

For such messages, assume the score is 206.835, which is what it would
be if we were to assume that `0 / 0.0` is `0` (which, apparently,
Wolfram Alpha does, which is good enough for me; for reference, it
considers `0 / 0` to be "indeterminate").

Change-Id: I0beb8012c64eb9a0edc412ae9c94c348c82ee190
Reviewed-on: https://gerrit.causes.com/26639
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>