Add stupidtable.js
I don't want to write my own table-sorter, nor do I want to pull in one
of the heavyweight "everything including the kitchen sink" solutions out
there; this is about the lightest I can find, although it is a little
flakey. It will have to do for now.

Downloaded from:

  https://github.com/joequery

Change-Id: I3d5448b5f528bd69d4e359e1a1798ebd9680ced0
Reviewed-on: https://gerrit.causes.com/26649
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>