Make /gerrit/authors sortable
As noted in previous commits, this plug-in is a little flakey but it
basically works. The number of data attributes required in the DOM is a
bit annoying, as is the need to carefully eliminate undefined and NaN
values. I might end up swapping this out for something built using
the jQuery sortElements plug-in at some point, which I suspect might be
more robust:

  https://github.com/padolsey/jQuery-Plugins/blob/master/sortElements/jquery.sortElements.js

Change-Id: I2befb3870c05a7989b5ea36e029ca29087d1d530
Reviewed-on: https://gerrit.causes.com/26651
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>