Show triangle on table headers to indicate sort order
By using a triangle character, ascending and descending sort order is
now visualized in table headers. I tried asking the internetz for the
correct icon/character to show when, but there seem to be general
confusion around this matter, so I just went with what one person wrote
on this SO thread:
http://stackoverflow.com/questions/338166/which-direction-should-the-arrows-point-in-a-sorted-table

"I don't think of them as arrows, but as a visual mnemonic of the
current state. So, showing a triangle pointing down shows descending
order. It is visually in line with the icon with the largest item (base
of the triangle) at the top of the list and the smallest (point of the
triangle) at the bottom."

Change-Id: I1d510b85bb285fc0961c9a134e36304dd884f593
Reviewed-on: https://gerrit.causes.com/26987
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>