Teach `#active?` to work with regular expressions
I will soon be adding a /gerrit/tags subpage, and so rather than adding
to the array of items passed into `#active?`, let's just teach it to
understand regexes.

This means that the array functionality is not being used any more;
that's ok. If it is still unused a week or two from now I'll actually
rip it out.

In updating the docs for the method, also corrected a minor inaccuracy
(it claimed a true return value, but it was merely truthy).

Related:

  https://xkcd.com/208/

Change-Id: I86c53d340256bf85fe60fee5a0c75e2ae2fddb8b
Reviewed-on: https://gerrit.causes.com/26991
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>