Use classes instead of ids for charts
This is a step towards flexibility and reusability. We're currently only
showing one chart per page, but we could conceivably want to show
multiple charts in the future.

This commit gets us a little closer to being able to do that, at low
cost. The styles, at least, are now completely re-usable, although the
JavaScript will need some more work if we want this stuff to appear more
than once on a given page.

Change-Id: If6a8c4ec0fdc94705e8c25924933662e2c2e15ba
Reviewed-on: https://gerrit.causes.com/27001
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>