Move stylesheets into "/assets" subdirectory
This is a step towards getting a real-ish "asset pipeline" in Sinatra.

I'd like to improve the organization of the JS in the web app, so I'll
be adding Sprockets in a subsequent commit. That will enable me to
compile JS into a single file from various sources.

Sass can already do something very similar for us, so I'm going to let
it keep doing its thing. The fact that we use Compass is a complicating
factor which makes it desirable to leave the Sass set-up as-is (I just
tried getting Sprockets to work with Sass + Compass in Sinatra and it is
a little painful; Sass without Compass is ok, but as soon as Compass is
in the mix things get tricky, so I'm going to leave the currently
working configuration in place).

Nevertheless, the very next step is going to be using Sprockets to
handle JS, so we may as well group all this stuff together under
"/assets" where it can be neatly organized and conceptually grouped
together.

Change-Id: I68dcef46595350f7055e95809ffdf92d092fc403
Reviewed-on: https://gerrit.causes.com/27003
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>