Pipe application JavaScript through Sprockets
With this commit we now have a mini "asset pipeline" for application
JavaScript. I'm not really concerned about minifying and concatenating
all the vendored code (asset-related performance is not really a problem
here; to make Twigg fast the lowest-hanging and juiciest fruit will be
adding caching[0]), but I do want to impose saner organization over the
non-vendored JavaScript in the app.

This commit adds Sprockets to handle the compilation, and moves our
application.js file into assets/javascripts/. In a subsequent commit
I'll start to break the JavaScript into multiple files, and then
reassemble them with Sprockets.

[0]: https://github.com/causes/twigg/issues/1

Change-Id: I16af0fce15560253efb66e6fe7c49f3f5ef72b92
Reviewed-on: https://gerrit.causes.com/27004
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>