Split application.js into separate files
To prevent application.js from becoming a random grab bag of unorganized
stuff, split its contents out into distinct files with well-defined
concerns.

Note that I chose to explicitly require the sub-files one-by-one rather
than using the `require_tree` directive, because I'll be wanting to
retain control over the order in which JS code is evaluated and
executed.

A subsequent commit will deal with a couple of places where we are using
in-template (footer) JS.

Change-Id: I2501c87026d9bb87a6d1402ac17356f1aae2228a
Reviewed-on: https://gerrit.causes.com/27005
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>