Load commit charts asynchronously
My initial intent here was to remove a couple of straggling references
to nvd3, but it became clear that the best way to do so would be to load
the chart data asynchronously rather than passing it into the partial
explicitly.

There is some duplication, particularly around the way progress is shown
and errors are displayed, but I'll address that in a future commit.

Note that until we implement caching[0], this will be slower than it
needs to be, because we effectively calculate most of the data twice.

[0]: https://github.com/causes/twigg/issues/1

Change-Id: Iabe58ec3c2a0371cb5a2a7ac20ca077249220630
Reviewed-on: https://gerrit.causes.com/27006
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>