Fix broken twiggrc template file
This commit fixes a syntax issue with the twiggrc example file related
to the "teams" setting. I ran into this bug when I tried to fire up the
project for the first time after copying the template to a ~/.twiggrc
file on my machine.

To avoid future problems, I added a simple spec for Twigg::Config that
verifies the correct behavior. To make testing slightly easier, I added
an option to the constructor of the Twigg::Config class that allows for
passing in a path to a twiggrc file as part of initialization.

Change-Id: I37ffa61e2ce37f3a1a1c396f3ec843d03f096e63
Reviewed-on: https://gerrit.causes.com/27012
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>