Move initialization of Twigg global var into application.js
We want this namespace to be set up and available from everywhere, so
let's prepare it first thing at the top of the application manifest.
Note that how this is done is a little counter-intuitive; we need to add
a `require_self` directive and make sure the code we want to run appears
after the directives.

Change-Id: I1af88d3a219c36b7d36a9dc240b497152f69687d
Reviewed-on: https://gerrit.causes.com/27016
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>