Convert russia.js to Backbone view
Backbone provides us with a nice means of structuring and organizing
client-side code. This is a more or less straight conversion of the old
code that was in russia.js to a new file at views/_russia.js.

I took inspiration from the declarative "data-view" pattern employed in
causes.git, but additionally I pass through any additional data
attributes on the element, which in this case I use to pass through
URLs. This makes life a little easier in the constructor function.

I took advantage of some of the niceties that Backbone provides, such as
easy binding of event handlers, scoped jQuery operations (via
`this.$()`), convenient access to `this.$el` and `this.el`, and
splitting the code into more readable functions (ie. `showNovelStats()`
and `render()`).

Change-Id: Ia117bcef2ff297aa4de76c2c71cf3b94bcb7c7f7
Reviewed-on: https://gerrit.causes.com/27116
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>