Port word cloud JS to Backbone
This is the minimal change required to the get the JS out of the bottom
of the /gerrit/tags template and into a Backbone view.

Behaviorally, it is identical to what was there before (ie. we haven't
yet implemented the per-author views, although this is important
groundwork for that) with one exception: we now pull in the stats JSON
via a separate AJAX request rather than inlining it, as it would be
cumbersome to pass it in via other means.

Change-Id: Iad6bfd68daf45ac441eec73c7b3f1c63d4f341ef
Reviewed-on: https://gerrit.causes.com/27201
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>