Add first rough cut of per-author word clouds
The scaling is a bit odd for people with few words in their clouds, the
tab contents are invisible until clicked on, and we need a better
zero-state (perhaps disable tabs for users without tags a given tab?),
but this is a working first cut.

Note that I had to set explicit height on the on the subviews, otherwise
Bootstrap's hidden content panes (width 0 by default) would cause the
word cloud plug-in to go into an infinite loop trying to place the
words.

Change-Id: I618691f2b621597dc58e7171698e2fd51588f5c7
Reviewed-on: https://gerrit.causes.com/27205
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>