Fix tag word cloud display for Kim
Kim's Gerrit authorname has multiple spaces in it ("Kim Douglas Clarence
Mason") and so our JS-side regex wasn't working when it came time to
generate IDs, and he wasn't getting a word cloud display.

This is evidence of how evil it is to have the logic existing both on
the server side (Ruby) and on the client. Nevertheless, for now, it is a
necessary evil.

Change-Id: Icd23b7a22015313d649f02405e3687179a194b0b
Reviewed-on: https://gerrit.causes.com/27208
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>