Add skeleton for twigg-cache gem
We're going to be adding a new gem to the Twigg family, twigg-cache. If
you want Twigg to go fast, install this gem and we'll cache results in a
Memcached instance for you.

We're doing it this way to avoid re-inventing the wheel (ie. if we were
to make a crude in-process cache store we'd have to re-implement expiry
semantics and so on).

Having at as a separate gem means that people without the means to run
Memcache can still use Twigg without bloating their dependency footprint
(this gem drags in the dalli gem as a dependency).

Change-Id: I027770547f8d2e32df6730d43af05207c97c3880
Reviewed-on: https://gerrit.causes.com/27302
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>