Make it possible to override Memcached max object size
The Memcached default is 1M, but some objects we would want to cache are
almost twice as large as that, even when compressed. For example, the
180-day /authors view at Causes currently has about 8,000 commits in its
commit set. This takes a little under 2M of memory when serialized and
compressed (about 237 bytes per commit).

Ideally we would break this down into smaller chunks and store them
separately, but for now, in the name of simplicity, I am just going to
make it possible pass through a value_max_value to the underlying Dalli
client. As most people likely won't want to set this, I don't provide a
default value.

In order for this to work, the `-I` option must be passed through to
Memcached at startup. For example, if Twigg is configured with a
value_max_bytes of 4194304, then Memcached should be started with `-I
4m`.

Change-Id: Ifa58c05cc6979ce0f9ae233f7846df4188003422
Reviewed-on: https://gerrit.causes.com/27309
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>