Teach `twigg app` -D/--daemon and -P/--pidfile options
In order to make the Twigg app easy to run as a service (eg. at boot
time) or monitor (eg. with monit), this commit teaches it some options
for going into the background as a daemon, and writing out a pidfile.

If a pidfile already exists, it bails. Note that we first employ a
race-prone check, before we daemonize, in order to be able to print a
diagnostic warning. After we've daemonized, we check again (trying to
obtain an exclusive lock) and we might die if we lose a race, at which
point it's too late to print out anything diagnostic because our stderr
and stdout have ben redirected to /dev/null.

Change-Id: I721aef639ad53c0584c44666fa11b925b25538bc
Reviewed-on: https://gerrit.causes.com/27578
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>