Bump version numbers for 0.0.3 release
Note the way this is currently set up it is most convenient if we
bump the version numbers of all the gems in lock-step with one another.
I might loosen this requirement in a future commit, to make it possible
to update just one of the gems and have the others track it via some
sort of "~> X.Y.Z" requirement.

Change-Id: I62647d7797619a1f143c7530bac7d823b7d3aa5e
Reviewed-on: https://gerrit.causes.com/27579
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>