Split CHANGELOG into separate files, in preparation for twigg-app 0.0.4
We're been doing something a little unorthodox here, in the name of
expediency, in that we have four different gems that are all being built
out of the same Git repo.

As there's only one repo, there can be only one "v0.0.1", "v0.0.2",
"v0.0.3" etc tag, unless we start "namespacing" the tags by including
the gem names in them.

For now we've side-stepped this question by releasing new versions of
each gem at the same time. Starting with the release of twigg-app
v0.0.4, which I am about to cut, we're going to do things more
granularly.

Specifically, we're going to "namespace" the tags, so the release will
be tagged with "twigg-app/v0.0.4", and we're going to release separate
updates for each gem only as needed.

This commit splits out the changelogs for each gem into separate files,
one in each subdirectory, and I'll go back and add the appropriate
namespaced tags to the previous release commits.

I've also bumped the version number on the twigg-app to 0.0.4.

Change-Id: I3004e24ba89cbae5880ce961cd52bade27802b2a
Reviewed-on: https://gerrit.causes.com/27603
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>