Change an `if !` to an `unless` for consistency
My general policy is to use `unless` for simple negative conditionals
rather than `if !`, and revert to `if ` only for complex/composite ones,
as they can be hard to read with `unless`.

In this file I wasn't applying this policy consistently, as I had:

  - unless <simple expression>
    (this is simple and negative, so should use `unless`; good)

  - if !<simple expression>
    (this is simple and negative, so should use `unless`; bad)

  - if !<subexpression> && !<subexpression>
    (complex/composite and negative, so should use `if`; good)

This commit fixes that second case above by turning it into an `unless`.

Change-Id: I24aa4c5782e5598f6e3a7648d257c90d5eee3910
Reviewed-on: https://gerrit.causes.com/27749
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>