Always work with UTF-8 for commit messages
I've observed mysterious failures to parse commit messages with UTF-8
characters in them on the CentOS box where I am running the Twigg app.
This same box handles the same repos just fine on the command line, but
explodes within the context of the Sinatra app.

Specifically, a commit message with UTF-8 in it (specifically, some em
dashes) was bailing when the parser tried to match a regex against a
line. Debugging statements revealed that the string contained multibyte
UTF-8 encoded codepoints, but Ruby was marking the string as
"ASCII-8BIT", and an invalid one at that.

Alas, due to our office move I am unable to troubleshoot this further
(the machine is down for the weekend), but I was able to establish that
the fix applied in this commit works. Specifically, the call to
`#force_encoding` fixes the problem; I added the `--encoding=UTF-8` arg
to `git log` just to be totally explicit and robust in the face of odd
local configuration or other unpredictable factors.

Maybe when the machine comes back up I'll be able to do more digging to
figure out what bizarre circumstance is forcing the log output to be
treated as though it weren't UTF-8.

Change-Id: I2116c4dc4a7ff466203a5ec3195e29a5f13cd7b5
Reviewed-on: https://gerrit.causes.com/28302
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>