Make twigg-cache gem degrade gracefully in the absence of memcache
If memcached goes down, behave as though everything is a cache miss
rather than exploding with a Dalli::RingError.

As there are only two methods (`#get` and `#set`) which require the
treatment, I did it the dumb and simple way. If we ever need to wrap
more methods like this I'll come up with something DRY-er.

Change-Id: Idad181702ff9616e7147dff26360087bc1462ab0
Reviewed-on: https://gerrit.causes.com/28303
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>