Bring all subdirectory Rakefiles up to DRY parity
Move pretty much everything up into common.rake, and ensure that each
subdirectory gem has the prereqs it needs (specifically RSpec) in order
to be able to run all the tasks.

Note that we're using instance variables here to communicate between the
Rakefiles and the common.rake tasks. This was deemed ever so slightly
less ugly than the alternatives (eg. $global vars, ENV vars).

Also note that the `import` statement works a little oddly. At the time
it is evaluated, it effectively acts as an "intent to include" rather
than an immediate evaluation of the contents of the file. Rake will
evaluate it only after it has fully evaluated the including file. This
means that you can't have anything in the including file that depends on
the immediate evaluation of the imported file (for example, constants
defined in the imported file).

This actually means that my intermediate step of moving RELEASE_PREREQS
up into common.rake was a misstep. The evaluation order is:

  - start evaluating Rakefile
  - register intent to evaluate common.rake
  - continue evaluating Rakefile
  - hit RELEASE_PREREQS, go boom

Change-Id: Iaed37bb30ceb2b70b3d83e693d4e650f91749ae0
Reviewed-on: https://gerrit.causes.com/28312
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>