Add twigg-pivotal gem
This is a first crude cut at implementing Pivotal Tracker integration.
With this commit, we have a `twigg pivotal stats` subcommand that prints
out a basic overview of the open stories in a Pivotal Tracker instance,
grouped by state (started, delivered etc).

The output is not too pretty, but I'll clean that up in subsequent
commits. The code should be fairly straightforward to plug in to the web
app, and I'll be doing that presently.

As with other elements in Twigg, this is being done as a separate gem in
order to keep the dependency footprint small; not everybody in the world
uses or needs Pivotal Tracker. I opted to use rest-client to talk to the
Pivotal API as the existing Pivotal gems haven't yet been updated for v5
of the API, and are probably overkill for what we need here.

Change-Id: I875a9e34570a8493aec07607e387cae6491959ad
Reviewed-on: https://gerrit.causes.com/28315
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>